Deleted WikiaHttpClient. Wikia APIs moved to separate folder.
