+++ src/Tests/Integration/wikia.integration.tests/WikiArticleListTests/AlphabeticalListTests.cs
@@ -3,6 +3,7 @@ using FluentAssertions;
 using NUnit.Framework;
 using wikia.Api;
 using wikia.Configuration;
+using wikia.Services;
 using wikia.tests.core;
 
 namespace wikia.integration.tests.WikiArticleListTests
