+++ src/Tests/Unit/wikia.unit.tests/WikiSearchSuggestionsTests/SuggestedPhrasesTests.cs
@@ -5,6 +5,7 @@ using System;
 using System.Threading.Tasks;
 using wikia.Api;
 using wikia.Models.SearchSuggestions;
+using wikia.Services;
 using wikia.tests.core;
 
 namespace wikia.unit.tests.WikiSearchSuggestionsTests
