+++ src/Wikia/Models/NavigationItem.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using wikia.Models.Mercury;
 
 namespace wikia.Models
 {
