+++ src/Wikia/wikia.csproj
@@ -27,7 +27,6 @@
 
   <ItemGroup>
     <Folder Include="Constants\" />
-    <Folder Include="Services\" />
   </ItemGroup>
 
 </Project>
